+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 # Head
 
 - Fixed: `host` options is now processed based `URL`.
+- Fixed: thrown error if `host` option being IP address.
 
 # 4.0.1 - 2016-10-27